Delete BundleHelper test project
