+++ src/installer/tests/Microsoft.NET.HostModel.Tests/AppHost.Bundle.Tests/HammerServiceTest.cs
@@ -4,7 +4,6 @@
 using System;
 using System.IO;
 using System.Linq;
-using BundleTests.Helpers;
 using Microsoft.DotNet.Cli.Build.Framework;
 using Microsoft.DotNet.CoreSetup.Test;
 using Microsoft.Extensions.DependencyModel;
@@ -51,7 +50,7 @@ namespace AppHost.Bundle.Tests
                 .CaptureStdErr()
                 .CaptureStdOut()
                 .EnableHostTracing()
-                .EnvironmentVariable(BundleHelper.CoreServicingEnvVariable, serviceBasePath)
+                .EnvironmentVariable(Constants.CoreServicing.EnvironmentVariable, serviceBasePath)
                 .Execute()
                 .Should()
                 .Pass()
