--- src/installer/tests/Microsoft.NET.HostModel.Tests/Helpers/BundleHelper.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Description>Bundle Test Helpers</Description>
-    <TargetFramework>$(TestInfraTargetFramework)</TargetFramework>
-    <AssemblyName>BundleHelper</AssemblyName>
-    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\TestUtils\TestUtils.csproj" />
-    <ProjectReference Include="..\..\..\managed\Microsoft.NET.HostModel\Microsoft.NET.HostModel.csproj" />
-  </ItemGroup>
-
-</Project>
