+++ src/installer/tests/Microsoft.NET.HostModel.Tests/Microsoft.NET.HostModel.Bundle.Tests/Microsoft.NET.HostModel.Bundle.Tests.csproj
@@ -14,7 +14,6 @@
   <ItemGroup>
     <ProjectReference Include="..\..\TestUtils\TestUtils.csproj" />
     <ProjectReference Include="..\..\..\managed\Microsoft.NET.HostModel\Microsoft.NET.HostModel.csproj" />
-    <ProjectReference Include="..\Helpers\BundleHelper.csproj" />
   </ItemGroup>
 
 </Project>
