+++ src/installer/tests/TestUtils/Constants.cs
@@ -59,6 +59,11 @@ namespace Microsoft.DotNet.CoreSetup.Test
             public const string EnvironmentVariable = "DOTNET_BUNDLE_EXTRACT_BASE_DIR";
         }
 
+        public static class CoreServicing
+        {
+            public const string EnvironmentVariable = "CORE_SERVICING";
+        }
+
         public static class DepsFile
         {
             public const string CommandLineArgument = "--depsfile";
